requirements.txt
numpy==1.22.1
opencv-python==4.5.5.64
pandas==1.4.0
Pillow==9.0.1
streamlit==1.9.0

